requirements.txt
asgiref==3.6.0
bleach==6.0.0
distlib==0.3.6
Django==3.2.19
django-debug-toolbar==4.1.0
django-extensions==3.2.3
django-multi-captcha-admin==2.0.0
django-ranged-response==0.2.0
django-robots==5.0
django-simple-captcha==0.5.18
django-summernote==0.8.20.0
django-taggit==4.0.0
filelock==3.12.2
Pillow==9.5.0
platformdirs==3.6.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
virtualenv==20.23.1
webencodings==0.5.1
